pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langchain-iceberg"
version = "0.1.0"
description = "LangChain integration for Apache Iceberg with native PyIceberg API support"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "LangChain Iceberg Contributors"}
]
keywords = ["langchain", "iceberg", "data-lake", "sql", "ai", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "langchain-core>=0.3.0",
    "pyiceberg>=0.7.0",
    "pydantic>=2.0",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
semantic = [
    "pyyaml>=6.0",
]
redis = [
    "redis>=5.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/langchain-ai/langchain-iceberg"
Documentation = "https://github.com/langchain-ai/langchain-iceberg#readme"
Repository = "https://github.com/langchain-ai/langchain-iceberg"
Issues = "https://github.com/langchain-ai/langchain-iceberg/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["langchain_iceberg*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=langchain_iceberg --cov-report=term-missing --cov-report=html"

